Resistance gain after administering drugs has been overhauled + more monitoring
Each drug now has a min and a max resistance gain to be applied to the bacteria when applied.
I have added monitors that show the lowest and highest resistance for each drug, as well as a plot.
A monitor has also been added to show exactly how much bacteria currently exist.